Fixing, little features and preparation. See desc.
Mregion canbyusedby fix
Checkbox fixed
Fixed debugui font dfont
Printtext scaling font available
added scale variable for bitmap font in element property
Preparation for advanced transparency bitmap features

+++ Engine2/Layer0/Vectors.h
@@ -37,8 +37,8 @@ public:
 		return (*this);
 	}
 
-	const int& w = x;
-    const int& h = y;
+	const float& w = x;
+    const float& h = y;
 
 	float DistanceFrom(Vector2 other) { 
 		return GetSquareDistance2(*this, other); }
